requirements.txt
streamlit
numpy
pandas
seaborn
matplotlib
folium
streamlit_folium
plotly.express
datetime
geopy
calendar